# packages/contracts/openapi/cirda-v1.yaml
openapi: 3.1.0
info:
  title: CIRDA Control Plane API
  version: 0.1.0
paths:
...
  /api/v1/ingest/events:
    post:
      tags:
      - ingest
      summary: Ingest Event
      operationId: ingest_event_api_v1_ingest_events_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEventRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/entities:
    get:
      tags:
      - entities
      summary: List Entities
      operationId: list_entities_api_v1_entities_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Limit
      - name: entity_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Type
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - entities
      summary: Create Entity
      operationId: create_entity_api_v1_entities_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/edges/{edge_id}:
    get:
      tags:
      - edges
      summary: Get Edge
      operationId: get_edge_api_v1_edges__edge_id__get
      parameters:
      - name: edge_id
        in: path
        required: true
        schema:
          type: string
          title: Edge Id
      - name: as_of
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - edges
      summary: Patch Edge Necessity
      operationId: patch_edge_necessity_api_v1_edges__edge_id__patch
      parameters:
      - name: edge_id
        in: path
        required: true
        schema:
          type: string
          title: Edge Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeNecessityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EdgeNecessityUpdate:
      properties:
        necessity:
          type: string
          title: Necessity
          description: 'Operator necessity annotation: required|optional|redundant|fallback|unknown'
      type: object
      required:
      - necessity
      title: EdgeNecessityUpdate
    EdgeResponse:
      properties:
        edge_id:
          type: string
          title: Edge Id
        source_id:
          type: string
          title: Source Id
        target_id:
          type: string
          title: Target Id
        relation:
          type: string
          title: Relation
        layer:
          type: string
          title: Layer
        confidence:
          type: number
          title: Confidence
        necessity:
          type: string
          title: Necessity
          default: unknown
        evidence_count:
          type: integer
          title: Evidence Count
          default: 0
      type: object
      required:
      - edge_id
      - source_id
      - target_id
      - relation
      - layer
      - confidence
      title: EdgeResponse
    EntityCreateRequest:
      properties:
        entity_id:
          type: string
          title: Entity Id
        entity_type:
          $ref: '#/components/schemas/EntityTypeSchema'
        name:
          type: string
          title: Name
        criticality:
          type: string
          title: Criticality
          default: unknown
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      additionalProperties: false
      type: object
      required:
      - entity_id
      - entity_type
      - name
      title: EntityCreateRequest
    EntityListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EntityResponse'
          type: array
          title: Items
        page:
          $ref: '#/components/schemas/PageMeta'
      additionalProperties: false
      type: object
      required:
      - items
      - page
      title: EntityListResponse
    EntityResponse:
      properties:
        entity_id:
          type: string
          title: Entity Id
        entity_type:
          anyOf:
          - $ref: '#/components/schemas/EntityTypeSchema'
          - type: string
          title: Entity Type
        name:
          type: string
          title: Name
        criticality:
          type: string
          title: Criticality
          default: unknown
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        aliases:
          items:
            type: string
          type: array
          title: Aliases
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - entity_id
      - entity_type
      - name
      title: EntityResponse
    EntityTypeSchema:
      type: string
      enum:
      - agent
      - tool
      - service
      - data
      - queue
      - credential
      - model
      title: EntityTypeSchema
    EvidenceEventResponse:
      properties:
        event_id:
          type: string
          title: Event Id
        source_id:
          type: string
          title: Source Id
        target_id:
          type: string
          title: Target Id
        relation:
          type: string
          title: Relation
        channel:
          type: string
          title: Channel
        observed_at:
          type: string
          format: date-time
          title: Observed At
        source_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Type
        target_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Type
        payload_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Payload Hash
        ingested_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ingested At
      additionalProperties: false
      type: object
      required:
      - event_id
      - source_id
      - target_id
      - relation
      - channel
      - observed_at
      title: EvidenceEventResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IngestEventRequest:
      properties:
        raw:
          additionalProperties: true
          type: object
          title: Raw
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
      additionalProperties: false
      type: object
      required:
      - raw
      title: IngestEventRequest
    IngestResult:
      properties:
        event:
          $ref: '#/components/schemas/EvidenceEventResponse'
        created:
          type: boolean
          title: Created
        edge_updated:
          type: boolean
          title: Edge Updated
      additionalProperties: false
      type: object
      required:
      - event
      - created
      - edge_updated
      title: IngestResult
...
    PageMeta:
      properties:
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      additionalProperties: false
      type: object
      required:
      - total
      - offset
      - limit
      title: PageMeta
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...